start working on doindex

+++ alcyon/orig/libsrc/qsort.c
@@ -35,8 +35,8 @@ static VOID _swap PROTO((char *a, char *b, size_t wid));
 
 int qsort(P(VOIDPTR) bas, P(size_t) num, P(size_t) siz, P(__compar_fn_t) cmp)
 PP(VOIDPTR bas;)
-PP(size_t num;)
-PP(size_t siz;)
+PP(int num;)
+PP(int siz;)
 PP(__compar_fn_t cmp;)
 {
 	register int i, j;
@@ -74,7 +74,7 @@ PP(__compar_fn_t cmp;)
 static VOID _swap(P(char *) a, P(char *) b, P(size_t) wid)
 PP(register char *a;)
 PP(register char *b;)
-PP(register size_t wid;)
+PP(register int wid;)
 {
 	register char tmp;
